fix: remove explicitly set binding.
I didn't document why I added that at the time, and there are also no
failing tests in `ash_postgres` (where we test this stuff). I do worry
that this may cause a problem with query building for someone down the
road, but we've made other fixes in this code that may have been a
"better" fix to whatever this weird code was originally *trying* to
fix.