fix: `has_many` with `limit` and `sort` not respected in `exists` query filters
Previously, the relationship limit was not being applied when using the relationship
in an `exists` query. After fixing that, then the limit was being applied in the
wrong place - it needs to be *inside* the exists check pre-filtering, instead
of outside it.

The test for this feature is in AshPostgres https://github.com/ash-project/ash_postgres/pull/667 -
this code makes that test pass.